mv NameList and LowerCaseDict to utils
